Add proper modularization to sox_effects
For better cohesion and coupling, move sox_effects related functions to sox_effects module.
Then perform initialization in __init__.py.

See the following for cohesion and coupling.
https://en.wikipedia.org/wiki/Coupling_(computer_programming)
