+++ torchaudio/__init__.py
@@ -1,4 +1,3 @@
-import atexit
 import os.path
 from pathlib import Path
 from typing import Any, Callable, Optional, Tuple, Union
@@ -20,6 +19,7 @@ from torchaudio._backend import (
 )
 from torchaudio._soundfile_backend import SignalInfo, EncodingInfo
 from torchaudio._internal import module_utils as _mod_utils
+from torchaudio.sox_effects import initialize_sox, shutdown_sox
 
 try:
     from .version import __version__, git_version  # noqa: F401
@@ -29,6 +29,7 @@ except ImportError:
 
 if _mod_utils.is_module_available('torchaudio._torchaudio'):
     from . import _torchaudio
+    initialize_sox()
 
 
 def load(filepath: Union[str, Path],
@@ -327,67 +328,6 @@ def get_sox_bool(i: int = 0) -> Any:
         return _torchaudio.sox_bool
     else:
         return _torchaudio.sox_bool(i)
-
-
-_SOX_INITIALIZED: Optional[bool] = False
-# This variable has a micro lifecycle. (False -> True -> None)
-# False: Not initialized
-# True: Initialized
-# None: Already shut down (should not be initialized again.)
-
-_SOX_SUCCESS_CODE = 0
-# defined at
-# https://fossies.org/dox/sox-14.4.2/sox_8h.html#a8e07e80cebeff3339265d89c387cea93a9ef2b87ec303edfe40751d9a85fadeeb
-
-
-@_mod_utils.requires_module("torchaudio._torchaudio")
-def initialize_sox() -> int:
-    """Initialize sox for use with effects chains.
-
-    You only need to call this function once to use SoX effects chains multiple times.
-    It is safe to call this function multiple times as long as ``shutdown_sox`` is not yet called.
-    Once ``shutdown_sox`` is called, you can no longer use SoX effects and calling this function
-    results in `RuntimeError`.
-
-    Note:
-        This function is not required for simple loading.
-
-    Returns:
-        int: Code corresponding to sox_error_t enum. See
-        https://fossies.org/dox/sox-14.4.2/sox_8h.html#a8e07e80cebeff3339265d89c387cea93
-    """
-    global _SOX_INITIALIZED
-    if _SOX_INITIALIZED is None:
-        raise RuntimeError('SoX effects chain has been already shut down. Can not initialize again.')
-    if not _SOX_INITIALIZED:
-        code = _torchaudio.initialize_sox()
-        if code == _SOX_SUCCESS_CODE:
-            _SOX_INITIALIZED = True
-            atexit.register(shutdown_sox)
-        return code
-    return _SOX_SUCCESS_CODE
-
-
-@_mod_utils.requires_module("torchaudio._torchaudio")
-def shutdown_sox() -> int:
-    """Showdown sox for effects chain.
-
-    You do not need to call this function as it will be called automatically
-    at the end of program execution, if ``initialize_sox`` was called.
-
-    It is safe to call this function multiple times.
-
-    Returns:
-        int: Code corresponding to sox_error_t enum. See
-        https://fossies.org/dox/sox-14.4.2/sox_8h.html#a8e07e80cebeff3339265d89c387cea93
-    """
-    global _SOX_INITIALIZED
-    if _SOX_INITIALIZED:
-        code = _torchaudio.shutdown_sox()
-        if code == _SOX_INITIALIZED:
-            _SOX_INITIALIZED = None
-        return code
-    return _SOX_SUCCESS_CODE
 
 
 def _audio_normalization(signal: Tensor, normalization: Union[bool, float, Callable]) -> None:
